requirements.txt
fastapi
uvicorn
streamlit
redis
google-generativeai
python-dotenv
pydantic
pydantic-settings
prometheus-client
aiohttp
pytest
httpx
pytest-cov
flake8
pytest-asyncio
black
flake8